:truck: [#23] feat: Move health check to checks file
